% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidsynth_fit.R
\name{evidsynth_fit}
\alias{evidsynth_fit}
\title{evidsynth_fit}
\usage{
evidsynth_fit(
  prevalence_data,
  progression_data,
  cut = TRUE,
  n_iter = 2000,
  n_burnin = 100,
  n_thin = 10
)
}
\arguments{
\item{cut}{Cut-point model to stop feedback; logical}

\item{n_iter}{Number of iterations}

\item{n_burnin}{Number of iterations in burn-in}

\item{n_thin}{Number of iterations to thin}
}
\description{
evidsynth_fit
}
