% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progression.R
\name{plot_progression}
\alias{plot_progression}
\title{Plot progression survival curve}
\usage{
plot_progression(evidsynth_obj)
}
\arguments{
\item{evidsynth_obj}{Output of model fit}
}
\value{
ggplot2 object
}
\description{
Plot progression survival curve
}
